Adopt the src/ layout and split the module by concern
The implementation was one 800-line levy/__init__.py mixing grid geometry,
interpolation, table I/O, parametrization algebra, the density functions,
fitting and sampling. This moves code and nothing else.

    src/levy/constants.py        grid geometry, fit bounds, parameter names
    src/levy/interpolation.py    Catmull-Rom interpolation, bound folding
    src/levy/tables.py           locating, loading, caching, repairing tables
    src/levy/parametrization.py  the five parametrizations, Parameters
    src/levy/distribution.py     levy(), neglog_levy()
    src/levy/fitting.py          fit_levy()
    src/levy/sampling.py         random()
    src/levy/_logging.py         the package logger
    src/levy/__init__.py         re-exports, __all__, __version__

levy/__init__.py re-exports the entire previous surface, so `import levy`
is unchanged for callers, and __all__ now states what is public -- before,
thirteen module globals were public by accident.

The src/ layout is the substantive part. Without it, `import levy` from the
repository root picks up the working tree instead of the installed package,
so tests can pass against code that was never packaged. That is exactly how
the unmerged `dev` branch moved the .npz files outside the package and
broke package_data resolution without anything noticing, and it is how my
own first clean-install check in an earlier PR silently tested the working
tree instead of the wheel.

Evidence that nothing moved numerically:

* all 280 goldens regenerate byte-identically, under --strict-bytes;
* scripts/compare_installs.py (added here) builds wheels from the previous
  commit and from this one, installs both, and diffs levy(), neglog_levy()
  and random() over 99,312 points as exact hex floats: every value matches
  bit for bit.

Two consequences worth knowing, both of which surfaced as test failures:

* Monkeypatching moves. data_dir() lives in levy.tables and calls the
  module-local user_cache_dir, so patching the re-exported levy.user_cache_dir
  no longer reaches it; patch levy.tables.user_cache_dir instead.
* Anything spawning a subprocess needs src/ on PYTHONPATH rather than the
  repository root.

Import graph is acyclic: _logging <- constants <- interpolation/tables <-
parametrization <- distribution <- fitting/sampling. pyproject, MANIFEST.in,
conftest and the two scripts follow the move.
