+++ src/levy/_logging.py
@@ -1,0 +1,32 @@
+# -*- encoding: utf-8 -*-
+#    Copyright (C) 2017 José M. Miotto
+#    This program is free software; you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation; either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with this program; if not, write to the Free Software
+#    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+
+"""The package logger.
+
+A library must not write to stdout. The NullHandler keeps it silent unless the
+application configures logging, per the stdlib logging guidance.
+"""
+
+import logging
+
+__all__ = ['logger']
+
+logger = logging.getLogger('levy')
+# Guarded: this module can be imported under a different name or
+# reloaded, and an unguarded add stacks a NullHandler per import on
+# the same logger object.
+if not any(isinstance(h, logging.NullHandler) for h in logger.handlers):
+    logger.addHandler(logging.NullHandler())
