+++ src/levy/data/manifest.json
